+++ gluoncv/auto/estimators/mask_rcnn/mask_rcnn.py
@@ -59,7 +59,7 @@ class MaskRCNNEstimator(BaseEstimator):
         The customized/default logger for this estimator.
     _logdir : str
         The temporary dir for logs.
-    _cfg : ConfigDict
+    _cfg : autocfg.dataclass
         The configurations.
     """
     def __init__(self, config, logger=None, reporter=None):
