+++ gluoncv/auto/estimators/ssd/ssd.py
@@ -55,7 +55,7 @@ class SSDEstimator(BaseEstimator):
         The customized/default logger for this estimator.
     _logdir : str
         The temporary dir for logs.
-    _cfg : ConfigDict
+    _cfg : autocfg.dataclass
         The configurations.
     """
     def __init__(self, config, logger=None, reporter=None):
