+++ gluoncv/auto/tasks/utils.py
@@ -11,6 +11,72 @@ from ..estimators import SSDEstimator, FasterRCNNEstimator, YOLOv3Estimator, Cen
 from ..estimators import ImageClassificationEstimator
 from .dataset import ObjectDetectionDataset
 
+
+class ConfigDict(dict):
+    """The view of a config dict where keys can be accessed like attribute, it also prevents
+    naive modifications to the key-values.
+
+    Parameters
+    ----------
+    config : dict
+        The sacred configuration dict.
+
+    Attributes
+    ----------
+    __dict__ : type
+        The internal config as a `__dict__`.
+
+    """
+    MARKER = object()
+    def __init__(self, value=None):
+        super(ConfigDict, self).__init__()
+        self.__dict__['_freeze'] = False
+        if value is None:
+            pass
+        elif isinstance(value, dict):
+            for key in value:
+                self.__setitem__(key, value[key])
+        else:
+            raise TypeError('expected dict, given {}'.format(type(value)))
+        self.freeze()
+
+    def freeze(self):
+        self.__dict__['_freeze'] = True
+
+    def is_frozen(self):
+        return self.__dict__['_freeze']
+
+    def unfreeze(self):
+        self.__dict__['_freeze'] = False
+
+    def __setitem__(self, key, value):
+        if self.__dict__.get('_freeze', False):
+            msg = ('You are trying to modify the config to "{}={}" after initialization, '
+                   ' this may result in unpredictable behaviour'.format(key, value))
+            warnings.warn(msg)
+        if isinstance(value, dict) and not isinstance(value, ConfigDict):
+            value = ConfigDict(value)
+        super(ConfigDict, self).__setitem__(key, value)
+
+    def __getitem__(self, key):
+        found = self.get(key, ConfigDict.MARKER)
+        if found is ConfigDict.MARKER:
+            if self.__dict__['_freeze']:
+                raise KeyError(key)
+            found = ConfigDict()
+            super(ConfigDict, self).__setitem__(key, found)
+        if isinstance(found, ConfigDict):
+            found.__dict__['_freeze'] = self.__dict__['_freeze']
+        return found
+
+    def __setstate__(self, state):
+        vars(self).update(state)
+
+    def __getstate__(self):
+        return vars(self)
+
+    __setattr__, __getattr__ = __setitem__, __getitem__
+
 def auto_suggest(config, estimator, logger):
     """
     Automatically suggest some hyperparameters based on the dataset statistics.
